Issue #792: separate inbound mapping processor actor into 2 actors.
