+++ services/connectivity/messaging/src/main/java/org/eclipse/ditto/services/connectivity/messaging/OutboundMappingProcessor.java
@@ -177,7 +177,7 @@ public final class OutboundMappingProcessor extends AbstractMappingProcessor<Out
                         .debug("Mapping <{}> produced <{}> messages.", mapper.getId(), messages.size());
 
                 if (messages.isEmpty()) {
-                    return Stream.of(MappingOutcome.dropped());
+                    return Stream.of(MappingOutcome.dropped(null));
                 } else {
                     return messages.stream()
                             .map(em -> {
@@ -189,18 +189,18 @@ public final class OutboundMappingProcessor extends AbstractMappingProcessor<Out
                                 final OutboundSignal.Mapped mapped =
                                         OutboundSignalFactory.newMappedOutboundSignal(outboundSignal, adaptable,
                                                 externalMessage);
-                                return MappingOutcome.mapped(mapped, adaptable.getTopicPath());
+                                return MappingOutcome.mapped(mapped, adaptable.getTopicPath(), null);
                             });
                 }
             } else {
                 logger.withCorrelationId(adaptable)
                         .debug("Not mapping message with mapper <{}> as MessageMapper conditions {} were not matched.",
                                 mapper.getId(), mapper.getIncomingConditions());
-                return Stream.of(MappingOutcome.dropped());
+                return Stream.of(MappingOutcome.dropped(null));
             }
         } catch (final Exception e) {
             return Stream.of(
-                    MappingOutcome.error(toDittoRuntimeException(e, mapper, adaptable), adaptable.getTopicPath())
+                    MappingOutcome.error(toDittoRuntimeException(e, mapper, adaptable), adaptable.getTopicPath(), null)
             );
         }
     }
